pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docker2k8s"
version = "1.0.0"
description = "Convert docker-compose.yml to production-ready Kubernetes manifests"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "docker2k8s contributors"}]
keywords = ["docker", "kubernetes", "docker-compose", "k8s", "devops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
docker2k8s = "docker2k8s.cli:main"

[tool.setuptools.packages.find]
include = ["docker2k8s*"]
